Refactor the liquity class to move statistics
